load_files/requirements.txt
google-cloud-bigquery>=2.1.0